requirements.txt
# API and Web Framework
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0

# Data Processing
numpy>=1.21.0
pandas>=1.3.0
biopython>=1.79

# API Clients
nvidia-api-client>=1.0.0
huggingface-hub>=0.16.4
requests>=2.26.0

# Database
sqlalchemy>=1.4.23
psycopg2-binary>=2.9.1
pymongo>=3.12.0

# Testing
pytest>=6.2.5
pytest-cov>=2.12.1

# Utilities
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.5 